requirements.txt
pygame-ce
pypubsub